requirements.txt
allure-pytest==2.12.0
httpx==0.23.3
jsonschema==4.17.3
pydantic==2.4.2
pydantic-settings==2.0.3
pytest==7.2.0
python-dotenv==0.21.0
pytest-asyncio==0.25.3
